Refactor: Move Regex test methods from GUIUtilTest to RegexValidatorTest in core
